% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ind.sat.derived.pm2.5.data.helper.R
\name{process_weather_raster}
\alias{process_weather_raster}
\title{Processes a yearly pollution raster}
\usage{
process_weather_raster(
  unit_raster_pol,
  unit_raster_pop,
  admin_level_shp_file,
  admin_level_shp_raster,
  var_type = "tmp.2m"
)
}
\arguments{
\item{admin_level_shp_file}{Shapefile representing the administrative boundaries.}

\item{admin_level_shp_raster}{Rasterized version of the administrative boundaries.}

\item{var_type}{Type of weather variable ("tmp.2m" for temperature, "tot.precip" for total precipitation).}

\item{unit_raster}{The raster representing the weather variable for a specific unit.}
}
\value{
A tibble containing processed weather data collapsed to the administrative level.
}
\description{
Processes a yearly pollution raster
}
