% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.derived.data.helper.R
\name{process_yearly_raw_pop_weighted_pol}
\alias{process_yearly_raw_pop_weighted_pol}
\title{Process Population-Weighted PM2.5 Pollution Data for a Given Year}
\usage{
process_yearly_raw_pop_weighted_pol(
  pol_raster_path,
  ref_admin_level_shp_file,
  ref_shp_file_to_be_rasterized,
  unit_pop_raster,
  unit_pop_raster_crp_msk
)
}
\arguments{
\item{pol_raster_path}{File path to the NetCDF (.nc) raster containing PM2.5 pollution data for a specific year.}

\item{ref_admin_level_shp_file}{A spatial object (e.g., \code{sf} or \code{sp}) representing the reference administrative boundaries for cropping and masking.}

\item{ref_shp_file_to_be_rasterized}{A spatial object with a column named \code{uid_for_rasterization} to be rasterized and used for aggregation.}

\item{unit_pop_raster}{A raster object representing the population data (e.g., at 1 km resolution).}

\item{unit_pop_raster_crp_msk}{A cropped and masked version of the population raster aligned with the reference shapefile.}
}
\value{
A data frame with administrative region identifiers (\code{sh_rast}), total population, and average population-weighted PM2.5 pollution
for the year extracted from the filename.
}
\description{
Computes average population-weighted PM2.5 pollution levels at a specified administrative level
using yearly pollution and population rasters, and a shapefile to rasterize for regional identification.
}
\details{
The function performs the following steps:
\enumerate{
\item Loads and preprocesses the pollution raster from NetCDF.
\item Crops and masks the pollution raster to the reference shapefile.
\item Resamples the pollution raster to match the resolution of the population raster.
\item Rasterizes the reference shapefile using \code{uid_for_rasterization}.
\item Combines pollution, population, and region rasters into a brick, converts to a data frame, and calculates population weights.
\item Computes population-weighted pollution averages by region.
\item Identifies and attempts to reprocess regions with missing or zero data.
\item Returns a final data frame with cleaned and complete population-weighted pollution estimates.
}
}
\seealso{
\code{\link{nc_to_raster_layer}}, \code{\link{matchResolution}}, \code{\link{fasterize}}, \code{\link{process_uncaptured_uids}}
}
