% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ind.sat.derived.pm2.5.data.helper.R
\name{chunk_list}
\alias{chunk_list}
\title{Function to chunk the list into groups of n, with the last sublist containing the remaining elements}
\usage{
chunk_list(lst, n)
}
\description{
Function to chunk the list into groups of n, with the last sublist containing the remaining elements
}
