% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.derived.data.helper.R
\name{nc_to_raster_layer}
\alias{nc_to_raster_layer}
\title{nc to rasterlayer}
\usage{
nc_to_raster_layer(nc_file_path, pm2.5_var_name, lat_var_name, long_var_name)
}
\description{
converts a .nc pollution file to a raster layer
}
