% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.derived.data.helper.R
\name{process_uncaptured_uids}
\alias{process_uncaptured_uids}
\title{Processes those polygons for which pollution and/or population data was not captured}
\usage{
process_uncaptured_uids(
  uids_not_captured,
  unit_pol_raster,
  unit_pop_raster,
  ref_shp_file_to_be_rasterized,
  resample_to_res = 0.001,
  res_resample_from = 0.00833333
)
}
\description{
Processes those polygons for which pollution and/or population data was not captured
}
