% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.derived.data.helper.R
\name{process_yearly_raw_pop_weighted_pol}
\alias{process_yearly_raw_pop_weighted_pol}
\title{process yearly pop weighted pol from raw rasters}
\usage{
process_yearly_raw_pop_weighted_pol(
  pol_raster_path,
  ref_admin_level_shp_file,
  ref_shp_file_to_be_rasterized,
  unit_pop_raster,
  unit_pop_raster_crp_msk
)
}
\description{
process yearly pop weighted pol from raw rasters
}
