% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ind.sat.derived.pm2.5.data.helper.R
\name{standardize_df_v2}
\alias{standardize_df_v2}
\title{Standardizes a data frame by taking care of all menial things (see description): v2}
\usage{
standardize_df_v2(df, matches_regex)
}
\arguments{
\item{df}{R dataframe/tibble to be standardized.}
}
\value{
returns a standardized dataframe. Please inspect once to make sure you get what is expected
        and continue further standardization, if needed.
}
\description{
Automatically coerces columns to right types, makes sure that all missing values in
character columns show up like "" and all missing values in numeric columns show up
like NA. But, it keeps the shrug uid columns in character form so as to not mess up
uid's that have a atleast one leading zeros.
}
\examples{

# standardize_df(df)

}
